+++ src/helpers/Drilldown.php
@@ -4,12 +4,15 @@ namespace venveo\characteristic\helpers;
 
 use Craft;
 use craft\base\Component;
+use craft\db\Query;
+use craft\db\Table;
 use craft\elements\db\ElementQueryInterface;
 use craft\helpers\ArrayHelper;
 use venveo\characteristic\Characteristic;
 use venveo\characteristic\elements\Characteristic as CharacteristicElement;
 use venveo\characteristic\elements\CharacteristicLinkBlock;
 use venveo\characteristic\elements\CharacteristicValue;
+use venveo\characteristic\elements\db\CharacteristicValueQuery;
 use venveo\characteristic\errors\CharacteristicGroupNotFoundException;
 use venveo\characteristic\models\CharacteristicGroup;
 
@@ -90,12 +93,23 @@ class Drilldown extends Component
         // Get the entire possible result set IDs
         $ids = $this->query->ids();
 
+        if (count($this->state->values)) {
+            $refined = clone($this->query);
+            $relations = ['and'];
+            foreach($this->state->values as $value) {
+                $relations[] = ['targetElement' => $value];
+            }
+            $refined->relatedTo($relations);
+            $ids = $refined->ids();
+        }
+
+
         // Get all link blocks that belong to the result set
         $linksQuery = CharacteristicLinkBlock::find()
             ->ownerId($ids);
 
-        // Get just the available characteristics in the result set
-        $linksQuery->indexBy('characteristicId');
+//        $values = CharacteristicValue::find()->relatedTo($linksQuery->ids())->all();
+//        Craft::dd(ArrayHelper::getColumn($values, '"value'));
 
         // Make sure we don't include characteristics we've already satisfied
         $skipIds = array_keys($this->state->satisfiedAttributes);
@@ -106,11 +120,34 @@ class Drilldown extends Component
         array_unshift($parsedSkipIds, 'and');
         $linksQuery->characteristicId($parsedSkipIds);
 
-        $blocks = $linksQuery->asArray()->all();
-        $characteristicIds = ArrayHelper::getColumn($blocks, 'characteristicId');
+        $blocks = $linksQuery->with(['values'])->all();
 
-        $characteristic = $this->_currentCharacteristic = CharacteristicElement::find()->groupId($this->_group->id)->id($characteristicIds)->orderBy('lft ASC')->one();
+        // We'll take a note of how many results each characteristic has and only suggest a characteristic that has
+        // at least 2 values
+        $indexedBlocks = [];
+        /** @var CharacteristicLinkBlock $block */
+        foreach($blocks as $block) {
+            $values = $block->getValues();
+            if ($values instanceof CharacteristicValueQuery) {
+                $valueIds = $values->ids();
+            } else {
+                $valueIds = ArrayHelper::getColumn($values, 'id');
+            }
+            if (!array_key_exists($block->characteristicId, $indexedBlocks)) {
+                $indexedBlocks[$block->characteristicId] = $valueIds;
+            } else {
+                $indexedBlocks[$block->characteristicId] = array_unique(array_merge($indexedBlocks[$block->characteristicId], $valueIds));
+            }
+        }
 
+        $allowedCharacteristics = [];
+        foreach($indexedBlocks as $characteristicId => $valueIds) {
+            if (count($valueIds) > 1) {
+                $allowedCharacteristics[] = $characteristicId;
+            }
+        }
+
+        $characteristic = $this->_currentCharacteristic = CharacteristicElement::find()->groupId($this->_group->id)->id($allowedCharacteristics)->orderBy('lft ASC')->one();
         return $characteristic;
     }